ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.26.0
scikit-learn==1.3.2
sentence-transformers==2.2.2 